+++ src/main/java/com/remind/api/mood/exception/MoodExceptionHandler.java
@@ -6,9 +6,9 @@ import com.remind.core.domain.common.exception.ActivityException;
 import com.remind.core.domain.common.exception.MemberException;
 import com.remind.core.domain.common.exception.MoodException;
 import com.remind.core.domain.common.response.ErrorResponse;
-import com.remind.core.domain.enums.ActivityErrorCode;
-import com.remind.core.domain.enums.MemberErrorCode;
-import com.remind.core.domain.enums.MoodErrorCode;
+import com.remind.core.domain.common.enums.ActivityErrorCode;
+import com.remind.core.domain.common.enums.MemberErrorCode;
+import com.remind.core.domain.common.enums.MoodErrorCode;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
