+++ src/main/java/com/remind/api/prescription/service/PrescriptionService.java
@@ -2,7 +2,6 @@ package com.remind.api.prescription.service;
 
 import com.remind.api.prescription.dto.PrescriptionDto;
 import com.remind.api.prescription.dto.request.CreatePrescriptionRequestDto;
-import com.remind.api.prescription.dto.request.PrescriptionInfoRequestDto;
 import com.remind.api.prescription.dto.response.CreatePrescriptionResponseDto;
 import com.remind.api.prescription.dto.response.PrescriptionInfoResponseDto;
 import com.remind.api.takingMedicine.service.TakingMedicineService;
@@ -11,9 +10,9 @@ import com.remind.core.domain.common.exception.MemberException;
 import com.remind.core.domain.common.exception.PrescriptionException;
 import com.remind.core.domain.connection.Connection;
 import com.remind.core.domain.connection.repository.ConnectionRepository;
-import com.remind.core.domain.enums.ConnectionErrorCode;
-import com.remind.core.domain.enums.MemberErrorCode;
-import com.remind.core.domain.enums.PresciptionErrorCode;
+import com.remind.core.domain.common.enums.ConnectionErrorCode;
+import com.remind.core.domain.common.enums.MemberErrorCode;
+import com.remind.core.domain.common.enums.PresciptionErrorCode;
 import com.remind.core.domain.member.Member;
 import com.remind.core.domain.member.enums.RolesType;
 import com.remind.core.domain.member.repository.MemberRepository;
