+++ src/main/java/com/remind/core/domain/common/enums/GlobalSuccessCode.java
@@ -1,4 +1,4 @@
-package com.remind.core.domain.enums;
+package com.remind.core.domain.common.enums;
 
 import lombok.Getter;
