+++ src/main/java/com/remind/core/domain/common/enums/SecurityErrorCode.java
@@ -1,19 +1,21 @@
-package com.remind.core.domain.enums;
+package com.remind.core.domain.common.enums;
 
 import com.remind.core.domain.common.response.ErrorResponse;
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
 
 @Getter
-public enum ActivityErrorCode implements BaseErrorCode {
+public enum SecurityErrorCode implements BaseErrorCode {
 
-    ACTIVITY_NOT_FOUND(404, "일치하는 activity가 없습니다.", HttpStatus.NOT_FOUND);
+
+    TOKEN_EXPIRED(401, "만료된 토큰입니다.", HttpStatus.UNAUTHORIZED),
+    INVALID_TOKEN(401, "유효하지 않은 토큰입니다.", HttpStatus.UNAUTHORIZED);
 
     private final int errorCode;
     private final String errorMessage;
     private final HttpStatus status;
 
-    ActivityErrorCode(int errorCode, String errorMessage, HttpStatus status) {
+    SecurityErrorCode(int errorCode, String errorMessage, HttpStatus status) {
         this.errorCode = errorCode;
         this.errorMessage = errorMessage;
         this.status = status;
