+++ src/main/java/com/remind/core/domain/common/exception/ConnectionException.java
@@ -1,7 +1,6 @@
 package com.remind.core.domain.common.exception;
 
-import com.remind.core.domain.enums.ConnectionErrorCode;
-import com.remind.core.domain.enums.PresciptionErrorCode;
+import com.remind.core.domain.common.enums.ConnectionErrorCode;
 import lombok.Getter;
 
 @Getter
