+++ src/main/java/com/remind/core/domain/common/exception/TakingMedicineException.java
@@ -1,7 +1,6 @@
 package com.remind.core.domain.common.exception;
 
-import com.remind.core.domain.enums.PresciptionErrorCode;
-import com.remind.core.domain.enums.TakingMedicineErrorCode;
+import com.remind.core.domain.common.enums.TakingMedicineErrorCode;
 import lombok.Getter;
 
 @Getter
